requirements.txt
requests==2.28.1
channels==3.0.5
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
psycopg2-binary==2.9.3
gunicorn==20.1.0
python-decouple==3.6